Rework saturating/edge behavior.
The example tool will now always "hit" the edge of the range on any
increase, unless it's _already_ at the edge of the range and --picky is
specified. Previously, --picky could keep it from reaching the edge of
the range.

You can now specify the bottom end of the range if you'd like it to be
something other than zero.

In the library, I've added conversion methods to/from percentage of
maximum.