Add gamma/exponent support.
By default the tool currently assumes your backlight's
control-to-brightness function is linear. This is basically never true,
because the control value is almost always controlling current (through
PWM) and LEDs have an inverse-exponential transfer function from current
to *perceived* brightness.

This adds a -e/--exponent switch to use a "gamma correction" mapping.
The correct value is usually around 3 or 4.